Move importer up in package
