+++ src/main/java/net/sf/jabref/logic/importer/ImportFormatReader.java
@@ -13,9 +13,9 @@ import net.sf.jabref.logic.importer.fileformat.BibTeXMLImporter;
 import net.sf.jabref.logic.importer.fileformat.BiblioscapeImporter;
 import net.sf.jabref.logic.importer.fileformat.BibtexImporter;
 import net.sf.jabref.logic.importer.fileformat.CopacImporter;
+import net.sf.jabref.logic.importer.fileformat.CustomImporter;
 import net.sf.jabref.logic.importer.fileformat.EndnoteImporter;
 import net.sf.jabref.logic.importer.fileformat.FreeCiteImporter;
-import net.sf.jabref.logic.importer.fileformat.Importer;
 import net.sf.jabref.logic.importer.fileformat.InspecImporter;
 import net.sf.jabref.logic.importer.fileformat.IsiImporter;
 import net.sf.jabref.logic.importer.fileformat.MedlineImporter;
