+++ src/main/java/net/sf/jabref/logic/importer/fileformat/MsBibImporter.java
@@ -7,6 +7,7 @@ import java.util.Objects;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 
+import net.sf.jabref.logic.importer.Importer;
 import net.sf.jabref.logic.importer.ParserResult;
 import net.sf.jabref.logic.msbib.MSBibDatabase;
 import net.sf.jabref.logic.util.FileExtensions;
