+++ src/test/java/net/sf/jabref/logic/openoffice/OOBibStyleTest.java
@@ -16,7 +16,7 @@ import java.util.Set;
 import net.sf.jabref.JabRefMain;
 import net.sf.jabref.logic.importer.ParserResult;
 import net.sf.jabref.logic.importer.fileformat.BibtexParser;
-import net.sf.jabref.logic.importer.fileformat.Importer;
+import net.sf.jabref.logic.importer.Importer;
 import net.sf.jabref.logic.journals.JournalAbbreviationLoader;
 import net.sf.jabref.logic.layout.Layout;
 import net.sf.jabref.logic.layout.LayoutFormatterPreferences;
